requirements.txt
numpy
Pillow
pytest
openai
